// src/interfaces/types/customizeModel.ts
export interface CustomizeModel {
  CornerRadius: number;
  BorderWeight: number;
  MainTitleFontSize: number;
  DescriptionFontSize: number;
  BarsTitleFontSize: number;
  BarsSubtitleFontSize: number;
  MainTitle: string; 
  Description: string;
  BarsPrimaryColor: string;
  BarsSubtitle: string;
  BarsBackground: string;
  BarsLabelBackground: string;
  BarSelectedBackground: string;
}